Rebased on frontend-dev
