+++ js-apps/meep-frontend/src/js/containers/exec/mobility-event-pane.js
@@ -22,21 +22,15 @@ import CancelApplyPair from '../../components/helper-components/cancel-apply-pai
 
 import {
   EXEC_EVT_MOB_TARGET,
-  EXEC_EVT_MOB_DEST
+  EXEC_EVT_MOB_DEST,
+  EVENT_CREATION_PANE_TABLE_LAYOUT,
+  EVENT_CREATION_PANE_LINE_LAYOUT
 } from '../../meep-constants';
 
 import {
   getElemFieldVal,
   FIELD_NAME
 } from '../../util/elem-utils';
-import { NC_TABLE_LAYOUT,  NC_LINE_LAYOUT} from '../../components/helper-components/nc-group';
-
-const LineLayout = (props) => {
-  return (<div>
-    
-
-  </div>);
-};
 
 class MobilityEventPane extends Component {
 
@@ -73,35 +67,6 @@ class MobilityEventPane extends Component {
 
   render() {
 
-    //check with list the target belongs to
-    if (this.values.eventTarget === undefined || this.values.eventTarget === '') {
-      return (
-        <div>
-          <Grid style={styles.field}>
-            <GridCell span="8">
-              <Select
-                style={styles.select}
-                label="Target"
-                outlined
-                options={_.map(this.props.MobTypes, elem => getElemFieldVal(elem, FIELD_NAME))}
-                onChange={(event)=>{this.values['eventTarget'] = event.target.value;}}
-                data-cy={EXEC_EVT_MOB_TARGET}
-              />
-            </GridCell>
-            <GridCell span="4">
-            </GridCell>
-          </Grid>
-
-          <CancelApplyPair
-            cancelText="Close"
-            applyText="Submit"
-            onCancel={this.props.onClose}
-            onApply={(e) => this.triggerEvent(e)}
-          />
-        </div>
-      );
-    }
-
     //let found = this.props.UEs.find(element => element.label == this.values.eventTarget);
     //find if its the selection was a UE, otherwise (in order) EDGE, FOG, EDGE-APP, UE-APP
     var target = this.values.eventTarget;
@@ -135,32 +100,81 @@ class MobilityEventPane extends Component {
         }
       }
     }
-    return (
-      <div>
+
+    const TargetSelectComponent = (
+      <>
+      <Select
+        style={styles.select}
+        label="Target"
+        outlined
+        options={_.map(this.props.MobTypes, elem => getElemFieldVal(elem, FIELD_NAME))}
+        onChange={(event)=>{this.values['eventTarget'] = event.target.value;}}
+        data-cy={EXEC_EVT_MOB_TARGET}
+      />
+      </>
+    );
+
+    const DestinationSelectComponent = (
+      <>
+      <Select
+        style= {styles.select}
+        label="Destination"
+        outlined
+        options={_.map(populateDestination, elem => getElemFieldVal(elem, FIELD_NAME))}
+        onChange={(event)=>{this.values['eventDestination'] = event.target.value;}}
+        data-cy={EXEC_EVT_MOB_DEST}
+      />
+      </>
+    );
+
+    const CancelApplyPairComponent = (
+      <>
+      <CancelApplyPair
+        cancelText="Close"
+        applyText="Submit"
+        onCancel={this.props.onClose}
+        onApply={(e) => this.triggerEvent(e)}
+      />
+      </>
+    );
+
+    //check with list the target belongs to
+    if (this.values.eventTarget === undefined || this.values.eventTarget === '') {
+      return (
+        <div>
+          <Grid style={styles.field}>
+            <GridCell span="8">
+              
+            </GridCell>
+            <GridCell span="4">
+            </GridCell>
+          </Grid>
+
+          <CancelApplyPair
+            cancelText="Close"
+            applyText="Submit"
+            onCancel={this.props.onClose}
+            onApply={(e) => this.triggerEvent(e)}
+          />
+        </div>
+      );
+    }
+
+    let Layout = null;
+    switch (this.props.layout) {
+    case EVENT_CREATION_PANE_TABLE_LAYOUT:
+      Layout = (
+        <>
         <Grid style={styles.field}>
           <GridCell span="8">
-            <Select
-              style={styles.select}
-              label="Target"
-              outlined
-              options={_.map(this.props.MobTypes, elem => getElemFieldVal(elem, FIELD_NAME))}
-              onChange={(event)=>{this.values['eventTarget'] = event.target.value;}}
-              data-cy={EXEC_EVT_MOB_TARGET}
-            />
+            <TargetSelectComponent />
           </GridCell>
           <GridCell span="4">
           </GridCell>
         </Grid>
         <Grid style={styles.block}>
           <GridCell span="8">
-            <Select
-              style= {styles.select}
-              label="Destination"
-              outlined
-              options={_.map(populateDestination, elem => getElemFieldVal(elem, FIELD_NAME))}
-              onChange={(event)=>{this.values['eventDestination'] = event.target.value;}}
-              data-cy={EXEC_EVT_MOB_DEST}
-            />
+            <DestinationSelectComponent />
           </GridCell>
           <GridCell span="4">
           </GridCell>
@@ -170,7 +184,7 @@ class MobilityEventPane extends Component {
       );
       break;
 
-    case NC_LINE_LAYOUT:
+    case EVENT_CREATION_PANE_LINE_LAYOUT:
       Layout = (
         <>
         <Grid style={styles.field}>
@@ -186,6 +200,7 @@ class MobilityEventPane extends Component {
       );
       break;
     }
+  
 
     return (
       <div>
